+++ nemo_aligner/algorithms/spin.py
@@ -32,6 +32,7 @@ from nemo_aligner.utils.ppo_utils import create_mask
 from nemo_aligner.utils.text_generation_utils import TrackLengthGPTModelTextGenerationStrategy
 from nemo_aligner.utils.train_utils import clip_gradients
 from nemo_aligner.utils.trainer_utils import check_progress, compute_limit_batches, compute_num_steps_per_epoch
+from nemo_aligner.utils.trt_llm import GPTGenerateTRTLLM
 from nemo_aligner.utils.utils import (
     batch_pad_to_fixed_len,
     clear_memory,
@@ -91,6 +92,7 @@ class SPINTrainer:
         logger,
         ckpt_callback,
         run_timer,
+        exp_manager,
     ):
         self.model = model
         self.train_dataloader = train_dataloader
@@ -135,6 +137,29 @@ class SPINTrainer:
         ), f"rollout_micro_batch_size [{self.model.cfg.spin.rollout_micro_batch_size}] must be a multiple of GBS [{self.model.cfg.global_batch_size}] // DP [{parallel_state.get_data_parallel_world_size()}]"
         self.rollout_micro_batch_size = self.model.cfg.spin.rollout_micro_batch_size
         assert self.rollout_micro_batch_size > 0, "`rollout_micro_batch_size` must be > 0"
+
+        self.use_trtllm_generation = self.cfg.trt_llm.get("enable", False) if "trt_llm" in self.cfg else False
+        if self.use_trtllm_generation:
+            # assert HAVE_TRTLLM, "TRTLLM generation was enabled but TRTLLM libraries could not be successfully imported"
+            self.trtllm_generate = GPTGenerateTRTLLM(
+                model_cfg=self.model.cfg,
+                end_strings=self.sampling_params["end_strings"],
+                tokenizer=self.model.tokenizer,
+                sample_temperature=self.sampling_params["temperature"],
+                sample_top_k=self.sampling_params["top_k"],
+                sample_top_p=self.sampling_params["top_p"],
+                repetition_penalty=self.sampling_params["repetition_penalty"],
+                max_generation_length=self.length_params["max_length"],
+                max_input_len=self.cfg.trt_llm.get(
+                    "max_input_len", self.model.cfg.encoder_seq_length - self.length_params["max_length"]
+                ),
+                generation_batch_size=self.model.cfg.spin.get("rollout_micro_batch_size", 4),
+                use_greedy=self.sampling_params.get("use_greedy", False),
+                trt_model_type=self.cfg.trt_llm.get("model_type", "gptnext"),
+                seed=self.model.cfg.get("seed", None),
+                unload_engine_train=self.cfg.trt_llm.get("unload_engine_train", False),
+                reshard_model=False,
+            )
 
         # for wandb table
         self.train_df = pd.DataFrame(columns=["step", "prompt", "response"])
@@ -205,7 +230,7 @@ class SPINTrainer:
 
         num_samples = 0
         gen_lengths = 0
-        num_samples += global_batch["actual"].shape[0]
+        num_samples += global_batch["chosen"].shape[0]
         gen_lengths += global_batch["generated_lengths"].sum()
         tensor_to_accumulate = torch.tensor(
             [gen_lengths, num_samples], dtype=torch.float32, device=torch.cuda.current_device(),
@@ -213,13 +238,16 @@ class SPINTrainer:
         torch.distributed.all_reduce(tensor_to_accumulate, group=parallel_state.get_data_parallel_group())
 
         (global_response_lengths, global_num_samples,) = tensor_to_accumulate.tolist()
-        metrics["avg_generated_lengths"] = global_response_lengths / global_num_samples
+        metrics["generated_lengths"] = global_response_lengths / global_num_samples
 
         return loss_mean, {**metrics, **trainer_metrics}
 
     @torch.no_grad()
     def get_generations(self, list_of_batches):
         self.model.prepare_for_inference()
+        if self.use_trtllm_generation:
+            # at this point self.model is the reference policy from cpu_weight_swap
+            self.trtllm_generate.refit(self.model)
 
         prompt_lengths = torch.cat([b["prompt_lengths"] for b in list_of_batches], dim=0)
         batch_max_length = prompt_lengths.max().item()
@@ -238,38 +266,46 @@ class SPINTrainer:
         )
         prompt_tokens = prompt_tokens.cuda(non_blocking=True)
         prompt_lengths = prompt_lengths.cuda(non_blocking=True)
+        inputs = (prompt_tokens, prompt_lengths)
 
-        strategy = TrackLengthGPTModelTextGenerationStrategy(
-            model=self.model, context_lengths=prompt_lengths, max_length=adj_generation_length
-        )
-        generations = self.model.generate(
-            inputs=(prompt_tokens, prompt_lengths),
-            length_params=self.length_params | {"max_length": adj_generation_length},
-            sampling_params=self.sampling_params,
-            strategy=strategy,
-        )
+        if self.use_trtllm_generation:
+            actor_output = self.trtllm_generate.generate(inputs)
+            response_tokens = actor_output["response_tokens"].cpu()
+            response_lengths = actor_output["response_lengths"].cpu()
+        else:
+            strategy = TrackLengthGPTModelTextGenerationStrategy(
+                model=self.model, context_lengths=prompt_lengths, max_length=adj_generation_length
+            )
+            generations = self.model.generate(
+                inputs=inputs,
+                length_params=self.length_params | {"max_length": adj_generation_length},
+                sampling_params=self.sampling_params,
+                strategy=strategy,
+            )
 
-        # this is a 1D LongTensor with the length of the responses where response is prompt+response
-        response_lengths = strategy.get_lengths().cpu()
-        max_response_length = response_lengths.max().item()
-        response_tokens = torch.LongTensor(generations["token_ids"]).cpu()
+            # this is a 1D LongTensor with the length of the responses where response is prompt+response
+            response_lengths = strategy.get_lengths().cpu()
+            max_response_length = response_lengths.max().item()
+            response_tokens = torch.LongTensor(generations["token_ids"]).cpu()
 
-        # Sanity check to validate response length.
-        if max_response_length != response_tokens.size(1):
-            # This may actually happen because NeMo does not always stop generation after `max_length` in batch mode
-            # => `response_tokens` may contain up to `max_length + max_context_length` tokens.
-            # TODO once NeMo fixes this issue we should be able to always raise an exception when the check above fails,
-            # and remove the `if` below.
-            if (
-                max_response_length >= response_tokens.size(1)
-                or response_tokens.size(1) != batch_max_length + adj_generation_length
-            ):
-                raise AssertionError(
-                    f"max response length ({max_response_length}) does not match the size of "
-                    f"`response_tokens` ({response_tokens.size(1)})"
-                )
+            # Sanity check to validate response length.
+            if max_response_length != response_tokens.size(1):
+                # This may actually happen because NeMo does not always stop generation after `max_length` in batch mode
+                # => `response_tokens` may contain up to `max_length + max_context_length` tokens.
+                # TODO once NeMo fixes this issue we should be able to always raise an exception when the check above fails,
+                # and remove the `if` below.
+                if (
+                    max_response_length >= response_tokens.size(1)
+                    or response_tokens.size(1) != batch_max_length + adj_generation_length
+                ):
+                    raise AssertionError(
+                        f"max response length ({max_response_length}) does not match the size of "
+                        f"`response_tokens` ({response_tokens.size(1)})"
+                    )
 
         self.model.finish_inference()
+        if self.use_trtllm_generation:
+            self.trtllm_generate.free()
 
         return response_tokens, response_lengths
 
@@ -362,12 +398,12 @@ class SPINTrainer:
 
                         # we update the pandas table here only during validation to avoid blowing up wandb storage space
                         # we update only for rank 0 although this is redudant because .log_table() only works on rank 0
-                        if torch.distributed.get_rank() == 0 and parallel_state.get_data_parallel_rank() == 0:
+                        if torch.distributed.get_rank() == 0:
                             self.train_df.loc[len(self.train_df)] = [
                                 self.step - 1,
                                 self.model.tokenizer.ids_to_text(global_batch["prompts_only"][0].tolist()),
                                 self.model.tokenizer.ids_to_text(
-                                    global_batch["generated"][0][
+                                    global_batch["rejected"][0][
                                         len(global_batch["prompts_only"][0]) : (
                                             len(global_batch["prompts_only"][0])
                                             + global_batch["generated_lengths"][0].item()
@@ -508,12 +544,12 @@ class SPINTrainer:
                         )
 
                     new_batch = {}
-                    new_batch["actual"] = act_tokens_pad
-                    new_batch["generated"] = gen_tokens_pad
+                    new_batch["chosen"] = act_tokens_pad
+                    new_batch["rejected"] = gen_tokens_pad
                     new_batch["attention_mask"] = attention_mask
                     new_batch["position_ids"] = position_ids
-                    new_batch["actual_mask"] = act_mask
-                    new_batch["generated_mask"] = gen_mask
+                    new_batch["chosen_mask"] = act_mask
+                    new_batch["rejected_mask"] = gen_mask
                     new_batch["prompts_only"] = batch["prompts_only"]
                     new_batch["generated_lengths"] = gen_lengths - batch["prompt_lengths"]
                     assert (gen_lengths - batch["prompt_lengths"] >= 0).all(), "negative generated length encountered"
@@ -521,8 +557,8 @@ class SPINTrainer:
                     logprobs = self.model.get_ref_policy_logprobs(new_batch).cpu()
                     act_logps, gen_logps = torch.split(logprobs, len(logprobs) // 2, dim=0)
 
-                    new_batch["ref_policy_log_probs_actual"] = act_logps
-                    new_batch["ref_policy_log_probs_generated"] = gen_logps
+                    new_batch["ref_policy_log_probs_chosen"] = act_logps
+                    new_batch["ref_policy_log_probs_rejected"] = gen_logps
 
                     start += batch_size