+++ nemo_aligner/utils/utils.py
@@ -28,7 +28,7 @@ from typing import Iterator, List
 from unittest.mock import patch
 
 import torch
-from megatron.core.dist_checkpointing.mapping import ShardedObject, ShardedTensorFactory
+from megatron.core.dist_checkpointing.mapping import LocalNonpersistentObject, ShardedObject, ShardedTensorFactory
 from megatron.core.num_microbatches_calculator import reconfigure_num_microbatches_calculator
 from omegaconf import DictConfig, OmegaConf
 from torch.masked import as_masked_tensor
@@ -512,6 +512,8 @@ def make_sharded_tensors_from_reference(reference_param, model_param, prefix: st
         return replace(reference_param, key=f"{prefix}.{reference_param.key}", data=model_param)
     if isinstance(reference_param, ShardedObject):
         return replace(reference_param, key=f"{prefix}.{reference_param.key}", data=model_param)
+    if isinstance(reference_param, LocalNonpersistentObject):
+        return LocalNonpersistentObject(model_param)
 
     assert (
         tuple(model_param.shape) == reference_param.local_shape